linebot_final_strict_reupload/requirements.txt
flask
line-bot-sdk==2.4.2
googletrans==4.0.0-rc1